Rename to dutch auction
