+++ contracts/DutchAuction.vy
@@ -1,91 +1,145 @@
 # @version ^0.3.7
 
-interface ERC20:
-    def transferFrom(asdf: address, sdlf:address, askdf: uint256) -> bool: view
+# vyper.interfaces.ERC20 does not include the mint and burn functions so we make our own interface
+interface ERC20WithAdminAccess:
+    def getBalanceOf(_address: address) -> uint256: nonpayable
+    def transferFrom(_from : address, _to : address, _value : uint256) -> bool: nonpayable
+    def approve(_spender : address, _value : uint256) -> bool: nonpayable
 
-ERC20Contract: public(ERC20)
+interface ERC721WithAdminAccess:
+    def ownerOf(_tokenId: uint256) -> address: nonpayable
+    def transferFrom(_from: address, _to: address, _tokenId: uint256): nonpayable
 
-event Buy:
+struct AuctionItem:
+    name: String[50] # See PublicGoods for a similar implementation of the names
     seller: address
-    buyer: address
-    price: uint256
+    startPrice: uint256
+    endPrice: uint256
+    startDate: uint256
+    endDate: uint256
+    nftTokenId: uint256
 
-#the amount of time that the auction will be active
-duration: public(uint256)
-#the seller of the auction item
-seller: public(address)
-#the starting price of the auction item
-startPrice: public(uint256)
-#the minimum price that the auction item can reach
-endPrice: public(uint256)
-#the time that the auction item was created & active
-startDate: public(uint256)
-#the final date when the auction will end no matter what
-endDate: public(uint256)
-#the address for the auction item NFT
-NFT: public(address)
-
+auctionItems: public(HashMap[String[50], AuctionItem])
+auctionItemsArr: public(DynArray[String[50], 100]) # A list of the names of all the auction items currently active
+erc20: ERC20WithAdminAccess
+erc721: ERC721WithAdminAccess
 
 @external
-def __init__(_duration: uint256, _startPrice: uint256, _endPrice: uint256, _NFT: address, Wolvercoin: address):
-    assert _endPrice >= _startPrice
-    self.duration = _duration
-    self.startPrice = _startPrice
-    self.endPrice = _endPrice
-    self.NFT = _NFT
-    self.seller = msg.sender
-    self.startDate = block.timestamp
-    self.endDate = self.startDate + self.duration
-    self.ERC20Contract = ERC20(Wolvercoin)
-
-@external
-def getEndDate() -> (uint256):
-    return self.endDate
+def __init__(erc20address: address, erc721address: address):
+    self.erc20 = ERC20WithAdminAccess(erc20address)
+    self.erc721 = ERC721WithAdminAccess(erc721address)
 
 @internal
-def transferFrom(_from : address, _to : address, _value : uint256):
-    return
+def findIndexOfItemInItemsArr(name: String[50]) -> int256:
+    for i in range(100):
+        if self.auctionItemsArr[i] == name:
+            return i
+    return -1
 
+# To call this function, one must approve the transfer via 721
 @external
-def _buy(buyer: address):
-    self.transferFrom(buyer, self.seller, self._getPrice())
-    self._endAuction()
+def createAuctionItem(name: String[50], startPrice: uint256, endPrice: uint256, startDate: uint256, endDate: uint256, nftTokenId: uint256):
+    assert name != ""
+    assert endPrice > 0
+    assert startPrice > endPrice
+    assert startDate >= block.timestamp
+    assert endDate > startDate
+    assert self.erc721.ownerOf(nftTokenId) == msg.sender
+
+    # Move the NFT to the property of this contract for safekeeping
+    self.erc721.transferFrom(msg.sender, self, nftTokenId)
+    
+    self.auctionItems[name] = AuctionItem({
+        name: name,
+        seller: msg.sender,
+        startPrice: startPrice,
+        endPrice: endPrice,
+        startDate: startDate,
+        endDate: endDate,
+        nftTokenId: nftTokenId
+    })
+    self.auctionItemsArr.append(name)
 
 @internal
-def _getPrice() -> (uint256):
-    assert self._isItemValid()
-    return self.startPrice - self.duration
+def _getPrice(name: String[50]) -> uint256:
+    assert name != ""
+    auctionItem: AuctionItem = self.auctionItems[name]
+    assert auctionItem.name == name
+    assert block.timestamp >= auctionItem.startDate
+    
+    if block.timestamp >= auctionItem.endDate:
+        return auctionItem.endPrice
 
-@internal
-def _isItemValid() -> (bool):
-    if self.endDate == 0:
-        return False
-    else:
-        return True
-
-@external
-def endAuction():
-    self.endDate = 0
-
-@internal 
-def _endAuction():
-#   self.startDate = null
-#   if block.timestamp > self.endDate:
-    self.endDate = 0
-#   if self.buy == False:
-#       if self._getPrice() < self.endPrice:
-#           self.startDate = 0
-#           self.endDate =0
-#       self.startDate = 0
-#   else:
-#       self.endDate = 0
+    # Linearly interpolate to find the price
+    timeSinceStart: uint256 = block.timestamp - auctionItem.startDate
+    totalDuration: uint256 = auctionItem.endDate - auctionItem.startDate
+    priceRange: uint256 = auctionItem.startPrice - auctionItem.endPrice
+    progress: decimal = convert(timeSinceStart, decimal) / convert(totalDuration, decimal)
+    return convert(convert(auctionItem.startPrice, decimal) - convert(priceRange, decimal) * progress, uint256)
 
 @external
-def isAuctionActive():
-    assert block.timestamp>= self.startDate
-    assert block.timestamp<self.endDate
+def getPrice(name: String[50]) -> uint256:
+    return self._getPrice(name)
+
+# To call this function, one must approve the monetary transfer via 20
+@external
+def buy(name: String[50]):
+    assert name != ""
+    auctionItem: AuctionItem = self.auctionItems[name]
+    assert auctionItem.name == name
+
+    assert block.timestamp > auctionItem.startDate
+    price: uint256 = self._getPrice(name)
+    assert self.erc20.getBalanceOf(msg.sender) >= price
+
+    self.erc20.transferFrom(msg.sender, auctionItem.seller, price)
+    self.erc721.transferFrom(self, msg.sender, auctionItem.nftTokenId)
+
+    self.auctionItems[name] = empty(AuctionItem)
+    i: int256 = self.findIndexOfItemInItemsArr(name)
+    if i != -1:
+        self.auctionItemsArr[i] = self.auctionItemsArr[len(self.auctionItemsArr) - 1] # Make the last element take the one you want to remove's place...
+        self.auctionItemsArr.pop() # ...and then remove the last element
 
 @external
-def getTimeRemaining() -> uint256:
-    #return self.endDate - block.timestamp
-    return 15
+def getActiveAuctionItems() -> DynArray[String[50], 100]:
+    return self.auctionItemsArr
+
+#region Trivial getters
+@external
+def getSeller(name: String[50]) -> address:
+    assert name != ""
+    auctionItem: AuctionItem = self.auctionItems[name]
+    assert auctionItem.name == name
+    return auctionItem.seller
+@external
+def getStartDate(name: String[50]) -> uint256:
+    assert name != ""
+    auctionItem: AuctionItem = self.auctionItems[name]
+    assert auctionItem.name == name
+    return auctionItem.startDate
+@external
+def getEndDate(name: String[50]) -> uint256:
+    assert name != ""
+    auctionItem: AuctionItem = self.auctionItems[name]
+    assert auctionItem.name == name
+    return auctionItem.endDate
+@external
+def getStartPrice(name: String[50]) -> uint256:
+    assert name != ""
+    auctionItem: AuctionItem = self.auctionItems[name]
+    assert auctionItem.name == name
+    return auctionItem.startPrice
+@external
+def getEndPrice(name: String[50]) -> uint256:
+    assert name != ""
+    auctionItem: AuctionItem = self.auctionItems[name]
+    assert auctionItem.name == name
+    return auctionItem.endPrice
+@external
+def getNFT(name: String[50]) -> uint256:
+    assert name != ""
+    auctionItem: AuctionItem = self.auctionItems[name]
+    assert auctionItem.name == name
+    return auctionItem.nftTokenId
+#endregion
